Add files via upload
A small project generating a derivative of tap code (where each letter has an assigned value in a "dictionary/table". Using a HackRF, transmitting that code on a frequency, sending it out in a translatable manner based on time. There are three key points to each letter. A preamble/declaration, a "section" declaration, and a "letter declaration".